Split 3 files exceeding 1000-line limit into submodules
- file_browser/tests.rs (1085→981): Extract snapshot and annotation
  tests into snapshot_tests.rs
- line_input/mod.rs (1035→950): Extract Message and Output enums
  into types.rs
- table/mod.rs (1001→785): Extract TableRow trait, Column, SortDirection,
  TableMessage, and TableOutput into types.rs
