+++ src/component/file_browser/mod.rs
@@ -914,4 +914,6 @@ impl Focusable for FileBrowser {
 }
 
 #[cfg(test)]
+mod snapshot_tests;
+#[cfg(test)]
 mod tests;
